+++ src/Specification/CompositeSpecification.php
@@ -65,23 +65,34 @@ class CompositeSpecification implements SpecificationInterface
         foreach ($this->specifications as $specification) {
             $rule = $specification->getRule();
             $parameters = $specification->getParameters();
+            $renamedParameters = [];
 
             foreach ($parameters as $parameterKey => $parameterValue) {
-                if (\array_key_exists($parameterKey, $processedParameters)) {
-                    // The parameter already exist in list. Add custom suffix.
-                    if (!\array_key_exists($parameterKey, $duplicatedParameters)) {
-                        $duplicatedParameters[$parameterKey] = 0;
-                    }
-
-                    $duplicatedParameters[$parameterKey]++;
-
-                    $modifiedParameterKey = $parameterKey.'_'.$duplicatedParameters[$parameterKey];
-                    $rule = \str_replace(':'.$parameterKey, ':'.$modifiedParameterKey, $rule);
-
-                    $processedParameters[$modifiedParameterKey] = $parameterValue;
-                } else {
-                    $processedParameters[$parameterKey] = $parameterValue;
+                if (!\array_key_exists($parameterKey, $processedParameters)) {
+                    continue;
                 }
+
+                // The parameter already exist in list. Add custom suffix, skipping names that are already taken.
+                do {
+                    $duplicatedParameters[$parameterKey] = ($duplicatedParameters[$parameterKey] ?? 0) + 1;
+                    $modifiedParameterKey = $parameterKey.'_'.$duplicatedParameters[$parameterKey];
+                } while (\array_key_exists($modifiedParameterKey, $processedParameters)
+                    || \array_key_exists($modifiedParameterKey, $parameters)
+                    || \in_array($modifiedParameterKey, $renamedParameters, true));
+
+                $renamedParameters[$parameterKey] = $modifiedParameterKey;
+            }
+
+            if (\count($renamedParameters)) {
+                // Rename whole parameter names in one pass: ":price" must not change ":price_min",
+                // and an already renamed parameter must not be renamed again.
+                $rule = \preg_replace_callback('/:([a-z_\d]+)/i', static function (array $match) use ($renamedParameters): string {
+                    return ':'.($renamedParameters[$match[1]] ?? $match[1]);
+                }, $rule);
+            }
+
+            foreach ($parameters as $parameterKey => $parameterValue) {
+                $processedParameters[$renamedParameters[$parameterKey] ?? $parameterKey] = $parameterValue;
             }
 
             $processedRules[] = $rule;